+++ tests/testit/test-list.R
@@ -6,9 +6,11 @@ assert('list_rmds() ignores thing in renv / packrat folder', {
   dir.create(dir, recursive = TRUE)
 
   dir.create(file.path(dir, 'renv'))
-  file.create(file.path(dir, 'renv/ignore.Rmd'))
+  file.create(rmd_renv <- file.path(dir, 'renv/ignore.Rmd'))
 
   rmd = list_rmds(dir)
+  (rmd %==% character())
+  rmd = list_rmds(files = rmd_renv)
   (rmd %==% character())
 
   unlink(dir, recursive = TRUE)